Move bitwise fuction from excellib to lib/engineering
