+++ .travis.yml
@@ -3,7 +3,7 @@ dist: xenial
 
 language: python
 
-python: ["3.5", "3.6", "3.7", "3.8"]
+python: ["3.6", "3.7", "3.8", "3.9"]
 
 sudo: false
