+++ src/pycel/excellib.py
@@ -681,52 +681,3 @@ def xsum(*args):
 
     # if no non numeric cells, return zero (is what excel does)
     return sum(data)
-
-
-@excel_math_func
-def bitand(op_x, op_y):
-    # Excel reference https://support.microsoft.com/en-us/office/
-    # bitand-function-8a2be3d7-91c3-4b48-9517-64548008563a
-    if op_x < 0 or op_y < 0:
-        return NUM_ERROR
-    return op_x & op_y
-
-
-@excel_math_func
-def bitor(op_x, op_y):
-    # Excel reference https://support.microsoft.com/en-us/office/
-    # bitor-function-f6ead5c8-5b98-4c9e-9053-8ad5234919b2
-    if op_x < 0 or op_y < 0:
-        return NUM_ERROR
-    return op_x | op_y
-
-
-@excel_math_func
-def bitxor(op_x, op_y):
-    # https://support.microsoft.com/en-us/office/
-    # bitxor-function-c81306a1-03f9-4e89-85ac-b86c3cba10e4
-    if op_x < 0 or op_y < 0:
-        return NUM_ERROR
-    return op_x ^ op_y
-
-
-@excel_math_func
-def bitlshift(number, pos):
-    # https://support.microsoft.com/en-us/office/
-    # bitlshift-function-c55bb27e-cacd-4c7c-b258-d80861a03c9c
-    if number < 0 or abs(pos) > 53 or number >= 2**48:
-        return NUM_ERROR
-    if pos < 0:
-        return bitrshift(number, abs(pos))
-    return number << pos
-
-
-@excel_math_func
-def bitrshift(number, pos):
-    # https://support.microsoft.com/en-us/office/
-    # bitrshift-function-274d6996-f42c-4743-abdb-4ff95351222c
-    if number < 0 or abs(pos) > 53 or number >= 2**48:
-        return NUM_ERROR
-    if pos < 0:
-        return bitlshift(number, abs(pos))
-    return number >> pos
